+++ include/parser.h
@@ -142,6 +142,8 @@ class Parser {
     std::unique_ptr<Statement> parseReturnStmt();
     std::unique_ptr<Statement> parseBreakStmt();
     std::unique_ptr<Statement> parseContinueStmt();
+    std::unique_ptr<Statement> parseJmpStmt();
+    std::unique_ptr<Statement> parseLabelStmt();
     std::unique_ptr<Statement> parseSwitchStmt();
     std::unique_ptr<Statement> parseCatchStmt();
     std::unique_ptr<Statement> parseThrowStmt();